Move dashboard to src/pages.
